rename search -> query and word -> term
This is meant to shorten function and variable names while also making
it more clear what is being operated on. The word "word" actually could
refer to multiple (quoted) words or even non-words (integer numbers for
example).

This uses "term" now instead. Also, the word "query" is used to refer to
the input from the user into the search box.

Finally, commonWordsFoundList was a very long var name. This shortens it
to commonTermHits, which still doesn't feel right, but it's a step in
the right direction.