fix negation terms (e.g. -term)
Previously it reset the score to 0, but the search would continue for
more terms on the page.

Now, if the negation is found on the page, it returns a score of 0 for
the page.

This also moves the search for negation first. Previously it searched
using the negation terms naively, as if they were normal terms. Only
after did it check if they were negation terms, and then searched all
over again.

Fix #15